【Fix PIR Unittest BUAA No. 13】Fix test_instance_norm_op
