+++ test/deprecated/legacy_test/CMakeLists.txt
@@ -677,7 +677,6 @@ set(TEST_CINN_OPS
     test_layer_norm_op_deprecated
     test_where_op
     test_arg_min_max_op
-    test_instance_norm_op
     test_instance_norm_op_deprecated)
 
 foreach(TEST_CINN_OP ${TEST_CINN_OPS})
