+++ test/legacy_test/CMakeLists.txt
@@ -1019,6 +1019,7 @@ set(TEST_CINN_OPS
     test_metrics
     test_meshgrid_op
     test_index_select_op
+    test_instance_norm_op
     test_mean_op
     test_clip_op
     test_gather_op
